feat: move config to top level script (#76)
* feat: delete rabbitmq_export's dummy test dir

* feat: move config to top level script

* feat: better logging

* feat: move terraform_drift_detection config to top level script

* feat: delete s3_prune_backups dummy test dir

* feat: move s3_prune_backups to top level script
