+++ .secrets.baseline
@@ -127,9 +127,9 @@
         "filename": "src/terraform_drift_detection/test/test_terraform.py",
         "hashed_secret": "aa891d4e101044bdd59c2b7ce36a2eb8f3b16793",
         "is_verified": false,
-        "line_number": 43
+        "line_number": 42
       }
     ]
   },
-  "generated_at": "2023-10-03T15:15:33Z"
+  "generated_at": "2023-11-29T20:17:56Z"
 }
