--- src/rabbitmq_export/test/test_export_definitions.py
@@ -1,2 +1,0 @@
-def test_placeholder():
-  pass
