--- src/s3_prune_backups/test/test_prune_backups.py
@@ -1,2 +1,0 @@
-def test_placeholder():
-  pass
